+++ src/services/SnaptchaService.php
@@ -69,17 +69,35 @@ class SnaptchaService extends Component
     }
 
     /**
+     * Returns the encoding type of the request.
+     */
+    public function getEncodingType(): string
+    {
+        $contentType = Craft::$app->getRequest()->getHeaders()->get('Content-Type');
+
+        return is_string($contentType) ? $contentType : '';
+    }
+
+    /**
      * Returns flattened posted values.
      */
     public function getPostedValues(): array
     {
-        $values = Craft::$app->request->getBodyParams();
+        $values = Craft::$app->getRequest()->getBodyParams();
 
         if (isset($values[Snaptcha::$plugin->settings->fieldName])) {
             unset($values[Snaptcha::$plugin->settings->fieldName]);
         }
 
         return $this->flattenValues($values);
+    }
+
+    /**
+     * Returns whether the request has a file upload.
+     */
+    public function hasFileUpload(): bool
+    {
+        return !empty($_FILES) && $_FILES['file']['error'] === UPLOAD_ERR_OK;
     }
 
     /**