+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Release Notes for Snaptcha
 
-## 5.0.0 - 2024-04-08
+## 5.1.0 - Unreleased
 
-### Added
+- Added an error message that is displayed if Snaptcha identifies a submission as spam when a file is uploaded, in which case the user must go back and resubmit their data ([#32](https://github.com/putyourlightson/craft-snaptcha/issues/32)).
+- Added the ability to maintain the encoding type of the submitted form on the error page.
+
+## 5.0.0 - 2024-04-08
 
 - Added compatibility with Craft 5.